fix(android/app): Increase intent priority for kmp download
